compression-service/requirements.txt
fastapi
python-multipart
Pillow
uvicorn[standard]
python-dotenv
